requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
pylint==2.3.1
six==1.12.0
slackclient==2.0.1
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.1
yarl==1.3.0
